v0.3.0: made rounding a bit more precise for Wand of Disintegration

+++ src/com/shatteredpixel/shatteredpixeldungeon/items/wands/WandOfDisintegration.java
@@ -90,7 +90,7 @@ public class WandOfDisintegration extends Wand {
 		
 		int lvl = level + chars.size() + terrainBonus;
 		int dmgMin = lvl;
-		int dmgMax = 8 + lvl * lvl / 3;
+		int dmgMax = (int) (8 + lvl * lvl / 3f);
 		for (Char ch : chars) {
 			ch.damage( Random.NormalIntRange( dmgMin, dmgMax ), this );
 			ch.sprite.centerEmitter().burst( PurpleParticle.BURST, Random.IntRange( 1, 2 ) );
